Cleanup PirateCheck
